pyproject.toml
[project]
name = "progressbar202-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedgen>=1.0.0",
    "pillow>=12.1.1",
]
